Extend FEInterface with kebab and pascal casevariants and moved template functions to ProtoUtil.hs
