+++ src/Ampersand/Prototype/GenAngularJSFrontend.hs
@@ -17,17 +17,11 @@ import Ampersand.Misc.HasClasses
 import Ampersand.Prototype.ProtoUtil
 import Ampersand.Runners (logLevel)
 import Ampersand.Types.Config
-import qualified RIO.List as L
 import qualified RIO.Text as T
 import System.FilePath
 import Text.StringTemplate
   ( StringTemplate,
-    Stringable,
-    checkTemplateDeep,
-    newSTMP,
-    render,
     setAttribute,
-    toString,
   )
 import Text.StringTemplate.GenericStandard ()
 
@@ -60,14 +54,6 @@ genRouteProvider fSpec ifcs = do
   mapM_ (logDebug . display) $ "Generated template: " : (map ("   " <>) . T.lines $ contents)
   writePrototypeAppFile "routeProvider.config.js" contents
   logDebug "Finish genRouteProvider."
-
-showTemplate :: Template -> [Text]
-showTemplate (Template a b) =
-  T.lines . T.intercalate "\n" $
-    ("Template (" <> T.pack b <> ")") :
-    map
-      ("  " <>)
-      [T.pack $ toString a]
 
 ------ Generate view html code
 isTopLevel :: A_Concept -> Bool
@@ -261,61 +247,3 @@ genControllerInterface fSpec interf = do
             . setAttribute "usedTemplate" controlerTemplateName
   let filename = "ifc" <> T.unpack (ifcName interf) <> ".controller.js"
   writePrototypeAppFile filename contents
-
------- Utility functions
--- data type to keep template and source file together for better errors
-data Template = Template (StringTemplate FilePath) FilePath
-
-readTemplate ::
-  (HasDirPrototype env) =>
-  FilePath ->
-  RIO env Template
-readTemplate templatePath = do
-  env <- ask
-  let absPath = getTemplateDir env </> templatePath
-  res <- readUTF8File absPath
-  case res of
-    Left err -> exitWith $ ReadFileError $ "Error while reading template." : err
-    Right cont -> return $ Template (newSTMP . T.unpack $ cont) absPath
-
-renderTemplate :: Maybe [TemplateKeyValue] -> Template -> (StringTemplate String -> StringTemplate String) -> Text
-renderTemplate userAtts (Template template absPath) setRuntimeAtts =
-  case checkTemplateDeep appliedTemplate of
-    -- BEWARE: checkTemplateDeep will hang if there are sub-attributes missing. I had such a case after I renamed ifcName
-    --         and ifcLabel to feiName and feiLabel. The template `routeProvider.config.js` needs those attributes in
-    --         for each interface provided.
-    ([], [], []) -> T.pack $ render appliedTemplate
-    (parseErrs@(_ : _), _, _) ->
-      templateError . T.concat $
-        [ T.pack $ "Parse error in " <> tmplt <> " " <> err <> "\n"
-          | (tmplt, err) <- parseErrs
-        ]
-    ([], attrs@(_ : _), _)
-      | isJust userAtts -> T.pack . render . fillInTheBlanks (L.nub attrs) $ appliedTemplate
-      | otherwise ->
-        templateError $
-          "The following attributes are expected by the template, but not supplied: " <> tshow attrs
-    ([], [], ts@(_ : _)) ->
-      templateError $
-        "Missing invoked templates: " <> tshow ts -- should not happen as we don't invoke templates
-  where
-    templateError msg =
-      exitWith $
-        ReadFileError
-          [ "*** TEMPLATE ERROR in:" <> T.pack absPath,
-            msg
-          ]
-    appliedTemplate = setRuntimeAtts . setUserAtts (fromMaybe [] userAtts) $ template
-    -- Set all attributes not specified to False
-    fillInTheBlanks :: [String] -> StringTemplate String -> StringTemplate String
-    fillInTheBlanks [] = id
-    fillInTheBlanks (h : tl) = setAttribute h False . fillInTheBlanks tl
-    setUserAtts :: [TemplateKeyValue] -> (StringTemplate String -> StringTemplate String)
-    setUserAtts kvPairs = foldl' fun id kvPairs
-      where
-        fun :: (Stringable b) => (StringTemplate b -> StringTemplate b) -> TemplateKeyValue -> (StringTemplate b -> StringTemplate b)
-        fun soFar keyVal = soFar . doAttribute keyVal
-        doAttribute :: (Stringable b) => TemplateKeyValue -> (StringTemplate b -> StringTemplate b)
-        doAttribute h = case tkval h of
-          Nothing -> setAttribute (T.unpack $ tkkey h) True
-          Just val -> setAttribute (T.unpack $ tkkey h) val
